requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Django==4.0.1
djangorestframework==3.13.1
generics==3.4.0
h11==0.13.0
idna==3.3
mysqlclient==2.1.0
outcome==1.1.0
Pillow==9.0.0
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.27.1
schedule==1.1.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
undetected-chromedriver==3.1.3
urllib3==1.26.8
websockets==10.1
wsproto==1.0.0
